Move to Module::Build to get windows working.
This means we don't need to run ./configure in monotonic_clock's
directory any more.

We kind of hack Module::Build by passing in direct C files in c_source,
even though the docs state that it's for directories. But the way they
are finding files seems to work, at the cost of them passing in the
source files to the compiler as include directories :-). Luckily those
are just warnings.