[03_oo.t] Add extra test for offset.
I might be wrong but it seems to me that the >0 test would only fail if the
math were backwards or if the timer wasn't running, but doesn't check that
the offset was actually added (since everything would be >0 in the normal
scheme of things). Adding a <0 test before the sleep test catches that case.

It does, however, add a little timing splinter: if the first $t2->now runs
more than 1ms after the initial call to ->new (for whatever reason), the
test will fail. I'm not sure I like that.